+++ tools/spec-gen-sdk/test/utils.test.ts
@@ -1,5 +1,5 @@
 import { parseSemverVersionString } from '../src/utils/parseSemverVersionString';
-import { removeAnsiEscapeCodes, removeDuplicatesFromRelatedFiles, diffStringArrays } from '../src/utils/utils';
+import { removeAnsiEscapeCodes, removeDuplicatesFromRelatedFiles, diffStringArrays, extractPathFromSpecConfig } from '../src/utils/utils';
 import { WorkflowContext } from '../src/automation/workflow';
 import * as typespecUtils from '../src/utils/typespecUtils';
 import { findMarkdownCodeBlocks, findSwaggerToSDKConfiguration } from '../src/utils/readme';
@@ -291,3 +291,33 @@ describe('find SDK Swagger Config from readme.md', () => {
   })
 
 })
+
+describe('extract and format the prefix from spec config path', () => {
+    it('should extract and format the prefix from tspConfigPath', () => {
+      const tspConfigPath = 'specification/myService.management/tspconfig.yaml';
+      const readmePath = undefined;
+      const result = extractPathFromSpecConfig(tspConfigPath, readmePath);
+      expect(result).toEqual('myservice-management');
+    });
+
+    it('should extract and format the prefix from readmePath', () => {
+      const tspConfigPath = undefined;
+      const readmePath = 'specification/myService/resource-manager/readme.md';
+      const result = extractPathFromSpecConfig(tspConfigPath, readmePath);
+      expect(result).toEqual('myservice');
+    });
+
+    it('should extract and format the prefix from readmePath', () => {
+      const tspConfigPath = undefined;
+      const readmePath = 'specification/myService/subservice/data-plane/readme.md';
+      const result = extractPathFromSpecConfig(tspConfigPath, readmePath);
+      expect(result).toEqual('myservice-subservice');
+    });
+
+    it('should return an empty string if paths are not provided', () => {
+      const tspConfigPath = undefined;
+      const readmePath = undefined;
+      const result = extractPathFromSpecConfig(tspConfigPath, readmePath);
+      expect(result).toEqual('');
+    });
+  });